dded two new classes to fit the PMT hit times in a similar way to the PMT charges.  WCSimTimePredictor predicts the time at which a photon will arrive at the PMT, taking into account the track vertex, direction and length (which it gets from the emission profiles) and WCSimTimeLikelihood2 converts this into a likelihood value based on the measured hit time and the known PMT time resolution.
GNUMakeFile: Add new time classes
default.cfg: Turned the time fit on by default
LinkDef: Add new time classes

ChargePredictor: Now has an EmissionProfiles class passed to it soit can share the same one as the TimePredictor
FitterTrackParMap: added the vertex time to the map
FitterTree: Add predicted and true time information to the HitComparison and tree
LikelihooDigit: Now stores the name of the PMT (e.g. 10inch_HQE)
LikelihoodDigitArray: When it creates a LikelihoodDigit is also passes it the PMT name
LikelihoodFitter: Changed a few things that didn't need printing, remembered to seed and update the track vertex time
LikelihoodTuner: Now has a PMTManager and a pointer to an EmissionProfiles instead of creating its own (so it can share the same one as the TimePredictor). Also added public function to get the cutoff distance for a track.  SolidAngle function now looks up the PMT radius instead of always using the same one
PhotonTrack & LikelihoodTrack: Made ::Print() do everything in one line instead of several
TotalLikelihood: Change WCSimTimeLikelihood to the new WCSimTimeLikelihood2, added GetPredictedTimeVector for the FitterTree to use, gave it a WCSimEmissionProfiles pointer to share between the charge and time predictors
TimeLikelihood: Added a ClassImp
TrackParameterEnums: Overloaded ostream so it prints out the name of the track type