Updates to the slicing and seeding.
Make sure that the seeding uses the results from the ring finder
and not just the vertex algorithm. This involved some changes
in the ring finding code such that it no longer stores a list
of all rings and isn't responsible for them.

The WCSimRecoEvent objects are now responsible for deleting their
attached WCSimRecoRing objects.