Get the expose height and QE for different PMT types, and the QE also as a function of distance travelled by the photon
DetectorParameters:
Add method to get PMT expose height so we can run with multiple PMT types.
Do the average QE vs distance to PMT because absorption is wavelength dependent

LikelihoodDigit:
To get PMT QE as a function of distance requires us to store a TGraph not a single number.

LikelihoodTuner:
Made QE a function of distance
Switched to use WCSim's version of the angular efficiency instead of MiniBooNE's